Update to upnp-1.8
Update master branch (1.8.x serie) to be able to install this version of
the library on the same machine than the "usual" 1.6.x pupnp library.
With these updates, the installation of upnp-1.8.x will not break
the existing applications that depend on the 1.6.x library.

Update:
 - include path from upnp to upnp-1.8
 - library name from libupnp* to libupnp-1.8
 - library name from libixml* to libixml-1.8
 - library name from libthreadutil* to libthreadutil-1.8
Rename libupnp.pc.in to libupnp-1.8.pc.in
Update libupnp.spec

Signed-off-by: Fabrice Fontaine <fabrice.fontaine@orange.com>