Adjust the library numbers for release
No need to adjust library numbers in configure.ac this time, because
this is the first release of branch 1.8.x and the library numbers have
already decent values.

A commit with this title updating the library numbers in configure.ac
is usually necessary right before tagging.

This is an empty commit.

Signed-off-by: Marcelo Roberto Jimenez <mroberto@users.sourceforge.net>