Drop -1.8 prefix at various places
This was introduced in

	e82c6bd6f014 ("Update to upnp-1.8")

to make it possible to install development files for both libupnp 1.6 and
libupnp 1.8. However having that suffix is a burden to programs linking
against libupnp. vlc developer Jean-Baptiste Kempf refused to update vlc
to use libupnp-1.8 with the suffix[1].

I didn't touch the library versions (LT_VERSION_*). There is no conflict
with the 1.6 branch so keeping the numbers should be ok.

[1] https://bugs.debian.org/868936#42