Fix large file system support
libupnp uses large file support (if available). If a program linking to
libupnp does not however it creates mismatches in callframes. See
Issue #51 for the results.

This simplifies LFS support by using AC_SYS_LARGEFILE_SENSITIVE instead of
manually defining _LARGE_FILE_SOURCE and _FILE_OFFSET_BITS (which is
useless on architectures where the size of off_t is fixed).

Furthermore additional logic is introduced to catch a library user without
64 bit wide off_t on such a platform.

upnp.h also makes use of off_t, but as this file includes FileInfo.h, the
latter is the single right place for this check.

This fixes #52 which is a generalized variant of #51.