Fix segmentation fault in http_MakeMessage
When upmpdcli is linked with libupnp 1.6.24 it fails with an
segmentation fault, but linking with libupnp worked just fine.

git bisect shows that commit 9c2e8ec8a0291ebe81959009e2f78edbdb47ced5 is
the bad one as the variable `extras` is not properly checked before
attempting to used it as a pointer.

Asure `extras` is not Null before using it.