As requested by Jean-Francois Dockes in github #49
in TemplateInclude.h in "replace_in_list()" (line 76). A parameter named
'new' is used. 'new' is a reserved keyword in c++ so this needs to be
replaced with 'newent' or whatever.

In httpreadwrite.c http_MakeMessage(), 2 cosmetic remarks: the 'E' case
is not part of the big ifelse. In practise this does not change
anything, but the reader loses time wondering why this is the case. Just
handle it like the other letters I think ? Also this saves a test
(maybe?) :=)

Also, it seems to me that there is no real warranty that the 'head'
value will always be not null (with possible future callers), so I'd
guard against it, like in 1.6 (I did read the code in the current caller
and it's ok, but a bit of defensive programming is often a good thing).