upnp/sample: Use UpnpResolveURL2 in samples
UpnpResolveURL is a tricky API to use correctly and the samples
allocate one bytes less than strictly required. This code has
been copied into other applications that similarly have problems
alocating the correct amount of space. Fortunately UpnpResolveURL2
is available and it avoids the need to allocate space in the caller
and does less copying too.