Allow the miniserver to listen on ULA or GUA IPv6 address
Currently, the miniserver only listens on the link local gIF_IPV6
address, this is an issue as the link local address is only available on
the current link (e.g. Ethernet) and not on the IPv6 "LAN". In other
words, this IPv6 address is not reachable if there is a router between
the UPnP device and the UPnP Control Point (which is the case for most
of our users).

So open a socket for gIF_IPV6_ULA_GUA to allow a device to call
UpnpRegisterRootDevice3 or UpnpRegisterRootDevice4 with the following
DescUrl:

sprintf(descDocUrlUlaGua, "http://[%s]:%d/%s", UpnpGetServerUlaGuaIp6Address(),
    UpnpGetServerUlaGuaPort6(), g_vars.descDocName);

Also add a new UpnpGetServerUlaGuaPort6 function to retrieve the port
for this ULA or GUA address (as it can be different from
UpnpGetServerPort6()).

Finally improve comments on those IPv6 functions in upnp.h and update
tv_device.c to make it IPv6 LLA or IPv6 LLA or GUA "aware".

Signed-off-by: Fabrice Fontaine <fontaine.fabrice@gmail.com>